loading,cambio collepsedSize
